+++ packages/Flow/src/package-test.ts
@@ -21,6 +21,7 @@ import './tests/execution-preview-tests';
 import './tests/viewer-tests';
 import './tests/column-picker-tests';
 import './tests/connect-interaction-tests';
+import './tests/string-list-tests';
 
 export let _package = new DG.Package();
 export { tests };